CB-5339 - Adding automated cloud backups to FreeIPA
This adds support for automated backup of the FreeIPA
service.  It backups up to the cloud provider defined
by the telemetry and uses the same logging bucket or
container as well as the Identity.

It will determine the backup directory root pathing
similarly to the way telemetry does.

It stores the backup configuration with the stack so
that as we add its own bucket later on, backward
compatibility can be maintained.